% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ED_defaults.R
\name{get_ED_default}
\alias{get_ED_default}
\title{Pull values from a given ED history.xml file
Assuming it is a history file from a defaults run}
\usage{
get_ED_default(xml, var)
}
\arguments{
\item{xml}{path to default history.xml}

\item{var}{variable of interest}
}
\value{
value
}
\description{
Pull values from a given ED history.xml file
Assuming it is a history file from a defaults run
}
