% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_species.R
\name{insert_species}
\alias{insert_species}
\title{Find nearby sites in BETY}
\usage{
insert_species(sp, case_accept_set, insert_PFT = FALSE, bety)
}
\value{
sp
}
\description{
Find nearby sites in BETY
}
