% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_traits.R
\name{insert_traits}
\alias{insert_traits}
\title{Insert traits in to the database ... we may already have this function. Who knows.}
\usage{
insert_traits(insert_dat, bety)
}
\arguments{
\item{insert_dat}{dataframe}
}
\value{
insert_dat dataframe
}
\description{
Insert traits in to the database ... we may already have this function. Who knows.
}
