% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearby_sites.R
\name{nearby_sites}
\alias{nearby_sites}
\title{Find nearby sites in BETY}
\usage{
nearby_sites(lat, lon, interval, bety = NA,
  config.php = "/fs/data3/ecowdery/pecan/web/config.php")
}
\value{
dataframe
}
\description{
Find nearby sites in BETY
}
