% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdistn.R
\name{pdistn}
\alias{pdistn}
\title{Given distribution details, make random samples}
\usage{
pdistn(xval, dist_params)
}
\value{
dataframe of distribution samples
}
\description{
Given distribution details, make random samples
}
