% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_VCD.R
\name{plot_VCD}
\alias{plot_VCD}
\title{Prepare VCD data and make plots}
\usage{
plot_VCD(model, var, keep.traits, plot, PDF,
  fpath = "/fs/data3/ecowdery/ED.Hydro/figures")
}
\description{
Prepare VCD data and make plots
}
