% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_VCD.R, R/plot_VDC.R
\name{plot_VDC}
\alias{plot_VDC}
\title{Prepare VDC data and make plots}
\usage{
plot_VDC(model, var, keep.traits, plot, PDF,
  fpath = "/fs/data3/ecowdery/ED.Hydro/figures")

plot_VDC(model, var, keep.traits, plot, PDF,
  fpath = "/fs/data3/ecowdery/ED.Hydro/figures")
}
\description{
Prepare VDC data and make plots

Prepare VDC data and make plots
}
