% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_add_samp_data.R
\name{prior_add_samp_data}
\alias{prior_add_samp_data}
\title{Add quartiles from sample data to the priors table}
\usage{
prior_add_samp_data(prior_in, samp)
}
\value{
priors dataframe
}
\description{
Add quartiles from sample data to the priors table
}
