% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_get_fit.R
\name{prior_get_fit}
\alias{prior_get_fit}
\title{Get a prior fit, specifically from a priors dataframe}
\usage{
prior_get_fit(prior_in, accepted_dists, plot = FALSE)
}
\arguments{
\item{prior_in}{one row of the priors dataframe}

\item{accepted_dists}{accepted dists}

\item{plot}{option to plot the fits}
}
\value{
fit object
}
\description{
Get a prior fit, specifically from a priors dataframe
}
