% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_input.R
\name{prior_input}
\alias{prior_input}
\title{Check if a prior is already in the database and enter it if it is not}
\usage{
prior_input(bety, variable_id.in, phylogeny.in, dist.name.in,
  parama.in = NA, paramb.in = NA, paramc.in = NA)
}
\arguments{
\item{bety}{BETY connection}

\item{variable_id.in}{BETY variable id}

\item{phylogeny.in}{BETY phylogeny}

\item{dist.name.in}{BETY distribution}
}
\value{
BETY prior id
}
\description{
Check if a prior is already in the database and enter it if it is not
}
