% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdistn.R
\name{rdistn}
\alias{rdistn}
\title{Given fit results from BETY, determine the distribution and sample}
\usage{
rdistn(fit, n = 1e+05)
}
\arguments{
\item{fit}{results from query to BETY priors table}
}
\value{
distribution sample
}
\description{
Given fit results from BETY, determine the distribution and sample
}
