% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Set1_color2hex.R
\name{Set1_color2hex}
\alias{Set1_color2hex}
\title{Simple function to get the hex code for my favorite colors}
\usage{
Set1_color2hex(color_in)
}
\description{
Simple function to get the hex code for my favorite colors
}
