% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_heatmap.R
\name{var_heatmap}
\alias{var_heatmap}
\title{Variance Heatmap}
\usage{
var_heatmap(PD.ALL, var_vector, full_range, keep.traits, color_range,
  barplot_facet = TRUE, title = "",
  fpath = "/fs/data3/ecowdery/ED.Hydro/figures/", plot.width,
  plot.height)
}
\description{
Variance Heatmap
}
