% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_keep_traits.R
\name{which_keep_traits}
\alias{which_keep_traits}
\title{Just an easy way to keep track of which traits we use in plots}
\usage{
which_keep_traits()
}
\description{
Just an easy way to keep track of which traits we use in plots
}
